More script tweaking

+++ cleanup.sh
@@ -1,41 +1,54 @@
 #!/usr/bin/env zsh
 
 # remove temporary files 
-ls | ggrep -P "^\d{2}_\d_\w\d{1,5}\.txt(\.bak)?$" | xargs rm
-
-touch words.list.tmp
+if [[ -f *.bak ]]
+then
+	rm *.bak
+fi
 
 for f in *.txt
 do
 	cat $f | aspell list >> words.list.tmp
 done
 
-if [[ -n words.list.tmp ]]
+if [[ -s words.list.tmp ]]
 then
 	awk '{ print "\\b" $1 "\\b" }' words.list.tmp > words.list
-	errors=$(ggrep -nf words.list *.txt)
+	$(ggrep -nf words.list *.txt) > errors
 	rm words.list
 fi
 rm words.list.tmp
 
-if [[ -n errors ]]
+if [[ -s errors ]]
 then
 	nvim -q errors
+fi
+
+if [[ -s errors ]]
+then
+	rm errors
 fi
 
 # final check
 for f in *.txt
 do
-	cat $f
-	echo '\n'
-done | less
+	cat $f >> check
+	echo >> check
+done
+
+less check
+rm check
 
 ls *.txt | pbcopy
 echo "Filenames copied to clipboard."
 
 PWD=$(pwd)
-rm -rf "${PWD}.bak"
-echo "Backup removed."
+if [[ -d "${PWD}.bak" ]]
+then
+	rm -rf "${PWD}.bak"
+	echo "Backup removed."
+fi
+
 mv $PWD $(dirname $(dirname $PWD))
 echo "${PWD} moved."
 cd ..
